refactor(klipper): extract KlippyServer business logic and fix audit issues
Separate KlippyServer (transport-agnostic business logic) from
KlippyUdsServer (thin UDS transport) and KlippyHttpServer (thin HTTP
transport). Both transports delegate to a shared KlippyServer instance
via callEndpoint(), eliminating business-logic duplication.

Audit fixes applied across all layers:
- Fix null deref in UDS handleRequest (crash on malformed JSON)
- Add mutex protection to SerialQueue (thread safety)
- Add retransmit limit to prevent infinite retry loops
- Complete state machine: Printing/Paused states + Error→Ready recovery
- Wire missing callbacks: job queue, history, filelist, power device
- Convert stubs to proper 501 error responses (service action, updates,
  log rollover, announcements, thumbnails, ZIP creation)
- Harden JWT: validate exp claim and reject "none" algorithm
- Fix path traversal using canonical path comparison
- Add oneshot token expiry (5-minute TTL)
- Replace deprecated gethostbyname() with getaddrinfo()
- Fix substr() overflow risks in /proc and URL parsing
- Fix DataDictionary compression buffer sizing (compressBound)
- Fix integer overflow in stepper clock and MotionTranslator
- Fix sequence number logic in KlipperDevice (first-block flag)
- Add 4-byte UTF-8 and surrogate pair handling in JSON parser
- Fix BedMeshObject to report actual mesh matrix data
- Fix probe object memory leak in KlippyInstance
- Add file size limit (16MB) to handleServerFilesGet
- Add HTTP error handling to Spoolman proxy (4xx/5xx)
- Fix CanTransport const_cast, PipeTransport fcntl error handling
- Add VLQ msgid overflow validation

Add 36 new audit tests covering state machine, callbacks, stubs,
security, UTF-8, and VLQ. All 1511 klipper + 39 HTTP tests pass.

+++ include/tether/klipper/motion/MotionTranslator.hpp
@@ -271,7 +271,7 @@ std::vector<AxisStepSequence> MotionTranslator<Dim, T>::translate(
                    (rawSteps[idx + mergedSamples].delta > 0 ? 1 : -1) == deltaSign) {
                 uint32_t nextAbs = static_cast<uint32_t>(
                     std::abs(rawSteps[idx + mergedSamples].delta));
-                if (static_cast<uint64_t>(totalSteps) + nextAbs > 65535) break;
+                if (static_cast<uint64_t>(totalSteps) + static_cast<uint64_t>(nextAbs) > 65535) break;
 
                 // Check if adding this sample would cause add to overflow.
                 if (mergedSamples > 1) {
